+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @dexto/core
 
+## 1.6.8
+
 ## 1.6.7
 
 ### Patch Changes