+++ packages/registry/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @dexto/registry
 
+## 1.6.8
+
 ## 1.6.7
 
 ### Patch Changes